requirements.txt
requests==2.25.1
beautifulsoup4==4.10.0
Pillow==8.4.0
selenium==4.0.0
